Move BraveNewsSubscriptions to a separate file
