+++ components/brave_news/browser/feed_v2_builder.h
@@ -14,7 +14,6 @@
 #include "base/functional/callback_forward.h"
 #include "base/memory/scoped_refptr.h"
 #include "base/memory/weak_ptr.h"
-#include "brave/components/brave_news/browser/brave_news_pref_manager.h"
 #include "brave/components/brave_news/browser/channels_controller.h"
 #include "brave/components/brave_news/browser/feed_fetcher.h"
 #include "brave/components/brave_news/browser/feed_sampling.h"
@@ -23,6 +22,7 @@
 #include "brave/components/brave_news/browser/suggestions_controller.h"
 #include "brave/components/brave_news/browser/topics_fetcher.h"
 #include "brave/components/brave_news/common/brave_news.mojom-forward.h"
+#include "brave/components/brave_news/common/subscriptions_snapshot.h"
 #include "components/history/core/browser/history_service.h"
 #include "mojo/public/cpp/bindings/pending_remote.h"
 #include "mojo/public/cpp/bindings/remote_set.h"
@@ -49,22 +49,22 @@ class FeedV2Builder {
 
   void AddListener(mojo::PendingRemote<mojom::FeedListener> listener);
 
-  void BuildFollowingFeed(const BraveNewsSubscriptions& subscriptions,
+  void BuildFollowingFeed(const SubscriptionsSnapshot& subscriptions,
                           BuildFeedCallback callback);
-  void BuildChannelFeed(const BraveNewsSubscriptions& subscriptions,
+  void BuildChannelFeed(const SubscriptionsSnapshot& subscriptions,
                         const std::string& channel,
                         BuildFeedCallback callback);
-  void BuildPublisherFeed(const BraveNewsSubscriptions& subscriptions,
+  void BuildPublisherFeed(const SubscriptionsSnapshot& subscriptions,
                           const std::string& publisher_id,
                           BuildFeedCallback callback);
-  void BuildAllFeed(const BraveNewsSubscriptions& subscriptions,
+  void BuildAllFeed(const SubscriptionsSnapshot& subscriptions,
                     BuildFeedCallback callback);
-  void EnsureFeedIsUpdating(const BraveNewsSubscriptions& subscriptions);
+  void EnsureFeedIsUpdating(const SubscriptionsSnapshot& subscriptions);
 
-  void GetSignals(const BraveNewsSubscriptions& subscriptions,
+  void GetSignals(const SubscriptionsSnapshot& subscriptions,
                   GetSignalsCallback callback);
 
-  void RecheckFeedHash(const BraveNewsSubscriptions& subscriptions);
+  void RecheckFeedHash(const SubscriptionsSnapshot& subscriptions);
 
  private:
   // FeedGenerator's will be called on a different thread. The data in
@@ -97,9 +97,9 @@ class FeedV2Builder {
   struct UpdateRequest {
     UpdateSettings settings;
     std::vector<UpdateCallback> callbacks;
-    BraveNewsSubscriptions subscriptions;
+    SubscriptionsSnapshot subscriptions;
 
-    explicit UpdateRequest(BraveNewsSubscriptions subscriptions,
+    explicit UpdateRequest(SubscriptionsSnapshot subscriptions,
                            UpdateSettings settings,
                            UpdateCallback callback);
     ~UpdateRequest();
@@ -125,7 +125,7 @@ class FeedV2Builder {
                                             PickArticles pick_article);
   static mojom::FeedV2Ptr GenerateAllFeed(FeedGenerationInfo info);
 
-  void UpdateData(const BraveNewsSubscriptions& subscriptions,
+  void UpdateData(const SubscriptionsSnapshot& subscriptions,
                   UpdateSettings settings,
                   UpdateCallback callback);
 
@@ -145,7 +145,7 @@ class FeedV2Builder {
 
   void NotifyUpdateCompleted();
 
-  void GenerateFeed(const BraveNewsSubscriptions& subscriptions,
+  void GenerateFeed(const SubscriptionsSnapshot& subscriptions,
                     UpdateSettings settings,
                     mojom::FeedV2TypePtr type,
                     FeedGenerator generator,
