+++ components/brave_news/browser/signal_calculator.h
@@ -12,7 +12,6 @@
 #include "base/containers/flat_map.h"
 #include "base/functional/callback_forward.h"
 #include "base/memory/weak_ptr.h"
-#include "brave/components/brave_news/browser/brave_news_pref_manager.h"
 #include "brave/components/brave_news/browser/channels_controller.h"
 #include "brave/components/brave_news/browser/feed_fetcher.h"
 #include "brave/components/brave_news/browser/publishers_controller.h"
@@ -28,6 +27,8 @@ using Signals =
     base::flat_map</*channel name or publisher_id*/ std::string, Signal>;
 using SignalsCallback = base::OnceCallback<void(Signals)>;
 
+class SubscriptionsSnapshot;
+
 class SignalCalculator {
  public:
   SignalCalculator(PublishersController& publishers_controller,
@@ -37,12 +38,12 @@ class SignalCalculator {
   SignalCalculator& operator=(const SignalCalculator&) = delete;
   ~SignalCalculator();
 
-  void GetSignals(const BraveNewsSubscriptions& subscriptions,
+  void GetSignals(const SubscriptionsSnapshot& subscriptions,
                   const FeedItems& feed,
                   SignalsCallback callback);
 
  private:
-  void OnGotHistory(const BraveNewsSubscriptions& subscriptions,
+  void OnGotHistory(const SubscriptionsSnapshot& subscriptions,
                     std::vector<mojom::FeedItemMetadataPtr> articles,
                     SignalsCallback callback,
                     history::QueryResults results);
